Move sentinel
