+++ zipline/pipeline/sentinels.py
@@ -1,0 +1,10 @@
+
+from zipline.utils.sentinel import sentinel
+
+
+NotSpecified = sentinel(
+    'NotSpecified',
+    'Singleton sentinel value used for Term defaults.',
+)
+
+NotSpecifiedType = type(NotSpecified)
